WIP on fixing a bug with client-side associations
Currently, if you create a new record, get one of its
has-many associations, and try to put another new record
into the ManyArray, the ManyArray will be reset after
calling `store.didCreateRecord` with the parent record.

This is a WIP commit on fixing some of the issues with
this scenario, although it will likely require deeper
architectural changes.